[CLEANUP]Moved jdbc package to conform w/ project
Renamed packages to conform with PDI DataService project
Removed unused classes
Retained usage of ThinUtil from core, since ThinCore is also used by the server
