--- src/main/java/org/pentaho/di/core/jdbc/FoundClause.java
@@ -1,41 +1,0 @@
-/*! ******************************************************************************
- *
- * Pentaho Data Integration
- *
- * Copyright (C) 2002-2014 by Pentaho : http://www.pentaho.com
- *
- *******************************************************************************
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- ******************************************************************************/
-
-package org.pentaho.di.core.jdbc;
-
-public class FoundClause {
-  private final String clause;
-  private final String rest;
-
-  public FoundClause( String clause, String rest ) {
-    this.clause = clause;
-    this.rest = rest;
-  }
-
-  public String getClause() {
-    return clause;
-  }
-
-  public String getRest() {
-    return rest;
-  }
-}
