+++ src/main/java/org/pentaho/di/trans/dataservice/jdbc/ThinConnection.java
@@ -20,7 +20,7 @@
  *
  ******************************************************************************/
 
-package org.pentaho.di.core.jdbc;
+package org.pentaho.di.trans.dataservice.jdbc;
 
 import java.sql.Array;
 import java.sql.Blob;
